# Conflicts:
#	backend/src/main/java/com/chattingweb/backend/controller/AdminController.java

+++ backend/src/main/java/com/chattingweb/backend/repository/admin/ReportRepository.java
@@ -3,5 +3,5 @@ package com.chattingweb.backend.repository.admin;
 import com.chattingweb.backend.entities.admin.Report;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReportRepository extends JpaRepository<Report, Integer> {
-}
+public interface ReportRepository extends JpaRepository<Report, Long> {
+  }
