+++ backend/src/main/java/com/chattingweb/backend/repository/conversation/MessageRepository.java
@@ -3,5 +3,5 @@ package com.chattingweb.backend.repository.conversation;
 import com.chattingweb.backend.entities.conversation.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MessageRepository extends JpaRepository<Message, Integer> {
-}
+public interface MessageRepository extends JpaRepository<Message, Long> {
+  }
